Move `BookmarksList` to `MainNavbarContainer`
